src/document-processor/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg2-binary==2.9.10
redis==5.2.1
boto3==1.36.4
